Bring back SlidingGuides and cycleButtonRadius, the ability to switch button radius in SettingsPageContents.
SlidingGuides is now part of ButtonArray, rather than a separate entity that is added to the instance of Stack in BasePage.

The issue in the previous version of this project, now defunct, where cycleButtonRadius caused a misalignment between SlidingGuides, ButtonArray and the movement/position of SettingsPageListTile is solved in this commit.