Fix odd behaviour at bottom right corner of SettingsPageListTile.
Essentially the problem was using a non-Aligned, rectangular BoxedContainer to build _FadingOverlay with a height equal to the height of SettingsPageList Tile. When adding a border to SettingsPageListTile instances of _FadingOverlay were shifted down by the border width causing it to extend beyond the curved boundary of each tile.

This commit adds settingsPageListTileBorderWidth to AppData, makes it a user-specified setting by adding cycleSettingsPageListTileBorderWidth and also makes it persist through app restarts.